requirements.txt
python-chess>=1.9.4
google-genai>=0.3.0
python-dotenv>=1.0.0
requests>=2.31.0
